refactor(email): give the transactional sender resolver one home
`pr-ready-email.ts` shipped a byte-for-byte copy of `dbos-digest.ts`'s
`resolveSender` — same body, same comment about reusing the invitation
provider. That's my duplication from #7159; extracting it now, before a
third caller copies it again.

It lives next to `createEmailSender`, which is what it builds. No
behavior change: both call sites resolve the same provider in the same
order, so a deployment with no provider still sends nothing and stamps
nothing.

The workflow-source snapshot moves because `dbos-digest.ts` changed.
Recovery-compatible — one call swapped for an identical one, no step
added, removed or reordered — so DBOS_WORKFLOW_VERSION stays at 11.
