+++ apps/api/src/notifications/dbos-digest.ts
@@ -26,9 +26,8 @@
 
 import { DBOS, SchedulerMode } from "@dbos-inc/dbos-sdk";
 import type { Kysely } from "kysely";
-import { getConfig } from "@/core/config";
 import { getBaseUrl } from "@/core/server-constants";
-import { createEmailSender, findEmailProvider } from "@/auth/email-providers";
+import { resolveTransactionalSender } from "@/auth/email-providers";
 import type { Database } from "@/storage/types";
 import { buildDigestEmail, type DigestRow } from "./digest-email";
 import { NotificationDataSchema } from "./schema";
@@ -165,22 +164,8 @@ function groupByRecipient(rows: PendingRow[]): PendingRow[][] {
   return [...byUser.values()];
 }
 
-/** Null when the deployment has no email provider — nothing is loaded, nothing
- *  is stamped, so the rows go out whenever one is configured.
- *
- *  The digest reuses the invitation provider — one configured sender, no new
- *  provider abstraction and no new env var. */
-function resolveSender() {
-  const auth = getConfig().auth;
-  const providers = auth.emailProviders ?? [];
-  const provider = auth.inviteEmailProviderId
-    ? findEmailProvider(providers, auth.inviteEmailProviderId)
-    : providers[0];
-  return provider ? createEmailSender(provider) : null;
-}
-
 async function sendOne(rows: PendingRow[]): Promise<void> {
-  const sender = resolveSender();
+  const sender = resolveTransactionalSender();
   if (!sender) throw new Error("no email provider configured");
   const { subject, html } = buildDigestEmail(rows, getBaseUrl());
   await sender({ to: rows[0]!.email, subject, html });
@@ -220,7 +205,7 @@ async function userDigestWorkflowFn(
 ): Promise<void> {
   const wait = dueAtMs - Date.now();
   if (wait > 0) await DBOS.sleep(wait);
-  if (!resolveSender()) return;
+  if (!resolveTransactionalSender()) return;
   const rows = await DBOS.runStep(() => loadPendingForUser(userId), {
     name: "loadPendingForUser",
   });
@@ -245,7 +230,7 @@ async function digestSweepWorkflowFn(
   const orphanedBefore = new Date(
     scheduledTime.getTime() - DEBOUNCE_MS - SWEEP_GRACE_MS,
   );
-  const pending = resolveSender()
+  const pending = resolveTransactionalSender()
     ? await DBOS.runStep(() => loadOrphaned(orphanedBefore), {
         name: "loadOrphanedNotifications",
       })
